updateBlogs page update confirm
