+++ src/Shared/Navbar/Navbar.jsx
@@ -12,7 +12,7 @@ import NavProfile from "./NavProfile";
 import useAuth from "../../Hooks/useAuth";
 import { CgGym } from "react-icons/cg";
 import { NotificationsMenu } from "./Notification";
-import Notification2 from "./Notification2";
+// import Notification2 from "./Notification2";
 
 function NavList() {
   const { user } = useAuth()
@@ -209,8 +209,8 @@ export function NavbarSimple() {
 
           <div className="flex gap-1 items-center">
             {/* Bell icon with notification button */}
-            {/* {user && <NotificationsMenu />} */}
-            {user && <Notification2 />}
+            {user && <NotificationsMenu />}
+            {/* {user && <Notification2 />} */}
 
             {/* User Profile component */}
             <span className="">
